# api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AeroLink Distributed Platform API",
    "version": "2.0.0",
    "description": "AeroLink Airline Systems Platform – Multi-Region Cloud API. Implements RESTful endpoints across 4 microservices: Booking (Aurora SQL), Baggage (DynamoDB NoSQL), Check-In, and Flight Operations. Secured with JWT/OAuth2 and RBAC.",
    "contact": { "name": "AeroLink Engineering", "email": "api@aerolink.io" },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local Development Gateway" },
    { "url": "https://api.aerolink.io", "description": "AWS Production (US-East-1)" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
  },
  "paths": {
...
    "/api/flights/schedule": {
      "get": {
        "summary": "Get full flight schedule",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Complete flight schedule with gate assignments and aircraft details" } }
      }
    },
    "/api/flights/pricing": {
      "get": {
        "summary": "Dynamic pricing matrix",
        "description": "Returns real-time pricing based on seat occupancy using tiered multipliers (simulates Lambda pricing engine).",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Dynamic price per flight with demand tiers" } }
      }
    },
    "/api/flights/update-status": {
      "post": {
        "summary": "Update flight status (Staff/Admin only)",
        "description": "Requires Staff or Admin role. Updates flight status and broadcasts FLIGHT_STATUS_UPDATED via SSE.",
        "tags": ["Flight Operations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["flightId", "status"],
                "properties": {
                  "flightId": { "type": "string", "example": "AL-101" },
                  "status": { "type": "string", "enum": ["ON_TIME", "DELAYED", "BOARDING", "DEPARTED", "ARRIVED", "CANCELLED"] },
                  "gate": { "type": "string", "example": "B14" },
                  "delayMins": { "type": "integer", "example": 45 },
                  "delayReason": { "type": "string", "example": "Weather" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated and broadcast" },
          "403": { "description": "Insufficient role (requires Staff or Admin)" }
        }
      }
    }
  }
}